bankroll-bot/requirements.txt
aiogram==3.6.0
SQLAlchemy==2.0.32
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.8.2
